Extract publication logic to buildSrc
After introducing `client-protocol` subproject, we need to publish
more than 1 project, so we extract the common logic to buildSrc.
