+++ subprojects/client-protocol/build.gradle.kts
@@ -1,6 +1,7 @@
 plugins {
     id("profiler.embedded-library")
     id("groovy")
+    id("profiler.publication")
 }
 
 dependencies {
